backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.23.0
sqlalchemy>=2.0.29
psycopg2-binary>=2.9.9
pydantic-settings>=2.2.1
email-validator>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
openai>=1.3.0
